requirements.txt
Flask>=3.0.0
nltk>=3.8.1
scikit-learn>=1.3.0
numpy>=1.24.0
PyPDF2>=3.0.0
python-docx>=1.0.0
Werkzeug>=3.0.0